fastapi/app/requirements.txt
fastapi==0.93.0
uvicorn[standard]==0.20.0
psycopg2-binary==2.9.5
pytest==7.2.2
httpx==0.23.3
bcrypt==4.0.1